added login functionality
